Modify sa_svp_buffer_copy and sa_svp_buffer_write to support multiple offsets

+++ reference/src/client/test/sa_svp_common.h
@@ -34,7 +34,7 @@ class SaSvpBufferAllocTest : public SaSvpBase {};
 
 class SaSvpBufferCheckTest : public ::testing::WithParamInterface<sa_digest_algorithm>, public SaSvpBase {};
 
-using SaSvpBufferTestType = std::tuple<size_t>;
+using SaSvpBufferTestType = std::tuple<long>;
 
 class SaSvpBufferCopyTest : public ::testing::WithParamInterface<SaSvpBufferTestType>, public SaSvpBase {};
